Change lttng-live plugin to use externally defined callbacks
Introduce lttng_live_process_ctf_event called for each CTF event.
Change lttng_live_should_quit to be application defined.
